# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jamila - Algeria Health & Beauty Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export interface Wilaya {
  id: number;
  code: number;
  name: string;
  nameAr: string;
  region: string;
  listingCount: number;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  nameAr: string;
  nameFr: string;
  icon: string;
  description: string;
  listingCount: number;
}

export type ListingPriceRange =
  (typeof ListingPriceRange)[keyof typeof ListingPriceRange];

export const ListingPriceRange = {
  budget: "budget",
  moderate: "moderate",
  premium: "premium",
  luxury: "luxury",
} as const;

export interface Listing {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  categorySlug: string;
  categoryName: string;
  wilayaId: number;
  wilayaName: string;
  wilayaCode: number;
  address: string;
  phone: string;
  email?: string | null;
  website?: string | null;
  description: string;
  rating: number;
  reviewCount: number;
  priceRange: ListingPriceRange;
  coverImage: string;
  images: string[];
  tags: string[];
  featured: boolean;
  verified: boolean;
  openingHours?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  createdAt: string;
}

export interface Service {
  id: number;
  name: string;
  nameAr?: string | null;
  description?: string | null;
  price?: number | null;
  priceMax?: number | null;
  duration?: string | null;
}

export interface Review {
  id: number;
  listingId: number;
  authorName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export type ListingDetail = Listing & {
  services: Service[];
  recentReviews: Review[];
};

export interface CreateReview {
  authorName: string;
  rating: number;
  comment: string;
}

export type CreateListingPriceRange =
  (typeof CreateListingPriceRange)[keyof typeof CreateListingPriceRange];

export const CreateListingPriceRange = {
  budget: "budget",
  moderate: "moderate",
  premium: "premium",
  luxury: "luxury",
} as const;

export interface CreateListing {
  name: string;
  categoryId: number;
  wilayaId: number;
  address: string;
  phone: string;
  email?: string | null;
  website?: string | null;
  description: string;
  priceRange: CreateListingPriceRange;
  coverImage: string;
  openingHours?: string | null;
}

export interface ListingsResponse {
  listings: Listing[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface SearchResponse {
  results: Listing[];
  total: number;
  query: string;
}

export interface MarketplaceStats {
  totalListings: number;
  totalCategories: number;
  totalWilayas: number;
  featuredCount: number;
  verifiedCount: number;
  topCategories: Category[];
  topWilayas: Wilaya[];
}

export type GetListingsParams = {
  category?: string;
  wilaya?: string;
  search?: string;
  featured?: boolean;
  minRating?: number;
  page?: number;
  limit?: number;
};

export type SearchListingsParams = {
  q: string;
  wilaya?: string;
  category?: string;
};