# lib/api-zod/src/generated/types/createListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jamila - Algeria Health & Beauty Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateListingPriceRange } from "./createListingPriceRange";

export interface CreateListing {
  name: string;
  categoryId: number;
  wilayaId: number;
  address: string;
  phone: string;
  email?: string | null;
  website?: string | null;
  description: string;
  priceRange: CreateListingPriceRange;
  coverImage: string;
  openingHours?: string | null;
}