# lib/api-zod/src/generated/types/listingDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jamila - Algeria Health & Beauty Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from "./listing";
import type { Review } from "./review";
import type { Service } from "./service";

export type ListingDetail = Listing & {
  services: Service[];
  recentReviews: Review[];
};