# lib/api-zod/src/generated/types/marketplaceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jamila - Algeria Health & Beauty Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { Wilaya } from "./wilaya";

export interface MarketplaceStats {
  totalListings: number;
  totalCategories: number;
  totalWilayas: number;
  featuredCount: number;
  verifiedCount: number;
  topCategories: Category[];
  topWilayas: Wilaya[];
}