# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jamila - Algeria Health & Beauty Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  listingId: number;
  authorName: string;
  rating: number;
  comment: string;
  createdAt: string;
}