# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jamila - Algeria Health & Beauty Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  nameAr?: string | null;
  description?: string | null;
  price?: number | null;
  priceMax?: number | null;
  duration?: string | null;
}