# lib/db/src/schema/services.ts
import { pgTable, text, integer, serial, real } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { listingsTable } from "./listings";

export const servicesTable = pgTable("services", {
  id: serial("id").primaryKey(),
  listingId: integer("listing_id").notNull().references(() => listingsTable.id),
  name: text("name").notNull(),
  nameAr: text("name_ar"),
  description: text("description"),
  price: real("price"),
  priceMax: real("price_max"),
  duration: text("duration"),
});

export const insertServiceSchema = createInsertSchema(servicesTable).omit({ id: true });
export type InsertService = z.infer<typeof insertServiceSchema>;
export type Service = typeof servicesTable.$inferSelect;